+++ client/loads/roomPage.cpp
@@ -15,10 +15,10 @@ void initBoxPlayers(rtype::ecs::Registry& reg)
     for (std::size_t i = 0; i < 4; i++) {
 
         prefabManager.createPrefab("white_square_" + std::to_string(i))
-            .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(265 * (i + 1), 350))
+            .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(265.0f * (static_cast<float>(i) + 1.0f), 350.0f))
             .addComponent<rtype::component::DebugColliderDisplay>(true)
-            .addComponent<rtype::component::Collider>(assetsManager.getTexture("white_square").width, assetsManager.getTexture("white_square").height)
-            .addComponent<rtype::component::Drawable>("white_square", 2, rtype::utils::Rectangle(0, 0, assetsManager.getTexture("white_square").width, assetsManager.getTexture("white_square").height), 1);
+            .addComponent<rtype::component::Collider>(static_cast<float>(assetsManager.getTexture("white_square").width), static_cast<float>(assetsManager.getTexture("white_square").height))
+            .addComponent<rtype::component::Drawable>("white_square", 2.0f, rtype::utils::Rectangle(0.0f, 0.0f, static_cast<float>(assetsManager.getTexture("white_square").width), static_cast<float>(assetsManager.getTexture("white_square").height)), 1);
     }
 }
 
@@ -27,13 +27,13 @@ void initTextRoom(rtype::ecs::Registry& reg)
     rtype::utils::PrefabManager& prefabManager = rtype::utils::PrefabManager::getInstance();
 
     prefabManager.createPrefab("room_id")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(0, 0))
-        .addComponent<rtype::component::Text>("ROOM ID:", rtype::component::Text::DEFAULT_FONT, 50, 3, raylib::WHITE)
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(0.0f, 0.0f))
+        .addComponent<rtype::component::Text>("ROOM ID:", rtype::component::Text::DEFAULT_FONT, 50, 3.0f, raylib::WHITE)
         .addComponent<rtype::component::Nameable>("ROOM ID");
 
     prefabManager.createPrefab("room_level")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(0, 100))
-        .addComponent<rtype::component::Text>("LEVEL: Loading...", rtype::component::Text::DEFAULT_FONT, 50, 3, raylib::WHITE)
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(0.0f, 100.0f))
+        .addComponent<rtype::component::Text>("LEVEL: Loading...", rtype::component::Text::DEFAULT_FONT, 50, 3.0f, raylib::WHITE)
         .addComponent<rtype::component::Nameable>("ROOM LEVEL");
 }
 
@@ -44,7 +44,7 @@ void initStartGameButton(rtype::ecs::Registry& reg)
 
     assetsManager.loadTexture("start_game_button", "assets/textures/buttons/StartButton.png");
     prefabManager.createPrefab("start_game_button")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(660, 600))
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(660.0f, 600.0f))
         .addComponent<rtype::component::Clickable>(
             [&]() {
                 rtype::network::message::client::StartGame startMessage = rtype::network::message::createEvent<rtype::network::message::client::StartGame>();
@@ -52,7 +52,7 @@ void initStartGameButton(rtype::ecs::Registry& reg)
                 rtype::network::Client::getInstance().getOutbox()->push(startPacked);
             })
         .addComponent<rtype::component::Collider>(assetsManager.getTexture("start_game_button").width, assetsManager.getTexture("login_start_button").height)
-        .addComponent<rtype::component::Drawable>("start_game_button", 1, rtype::utils::Rectangle(0, 0, assetsManager.getTexture("login_start_button").width, assetsManager.getTexture("login_start_button").height), 1);
+        .addComponent<rtype::component::Drawable>("start_game_button", 1.0f, rtype::utils::Rectangle(0.0f, 0.0f, assetsManager.getTexture("login_start_button").width, assetsManager.getTexture("login_start_button").height), 1);
 }
 
 void initSpaceShip(rtype::ecs::Registry& reg)
@@ -62,20 +62,20 @@ void initSpaceShip(rtype::ecs::Registry& reg)
 
     assetsManager.loadTexture("space_ship", "assets/textures/ships.png");
     prefabManager.createPrefab("space_ship_1")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(285, 410))
-        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1, 3, 32, 16), 2)
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(285.0f, 410.0f))
+        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1.0f, 3.0f, 32.0f, 16.0f), 2)
         .addComponent<rtype::component::Nameable>("SPACE SHIP 1");
     prefabManager.createPrefab("space_ship_2")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(550, 410))
-        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1, 20, 32, 16), 2)
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(550.0f, 410.0f))
+        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1.0f, 20.0f, 32.0f, 16.0f), 2)
         .addComponent<rtype::component::Nameable>("SPACE SHIP 2");
     prefabManager.createPrefab("space_ship_3")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(815, 410))
-        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1, 37, 32, 16), 2)
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(815.0f, 410.0f))
+        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1.0f, 37.0f, 32.0f, 16.0f), 2)
         .addComponent<rtype::component::Nameable>("SPACE SHIP 3");
     prefabManager.createPrefab("space_ship_4")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(1080, 410))
-        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1, 54, 32, 16), 2)
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(1080.0f, 410.0f))
+        .addComponent<rtype::component::Drawable>("space_ship", 0, rtype::utils::Rectangle(1.0f, 54.0f, 32.0f, 16.0f), 2)
         .addComponent<rtype::component::Nameable>("SPACE SHIP 4");
 }
 
@@ -86,14 +86,14 @@ void initBackRoomButton(rtype::ecs::Registry& reg)
     assetsManager.loadTexture("back_room_button", "assets/textures/buttons/BackButton.png");
 
     prefabManager.createPrefab("back_room_button")
-        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(685, 700))
+        .addComponent<rtype::component::Transform>(rtype::utils::Vector<float>(685.0f, 700.0f))
         .addComponent<rtype::component::Clickable>(
             [&]() {
                 rtype::utils::SceneManager& sceneManager = rtype::utils::SceneManager::getInstance();
                 sceneManager.setNextScene(sceneManager.getPreviousScene());
             })
         .addComponent<rtype::component::Collider>(assetsManager.getTexture("back_room_button").width, assetsManager.getTexture("back_room_button").height)
-        .addComponent<rtype::component::Drawable>("back_room_button", 1, rtype::utils::Rectangle(0, 0, assetsManager.getTexture("back_room_button").width, assetsManager.getTexture("back_room_button").height), 1);
+        .addComponent<rtype::component::Drawable>("back_room_button", 1.0f, rtype::utils::Rectangle(0.0f, 0.0f, assetsManager.getTexture("back_room_button").width, assetsManager.getTexture("back_room_button").height), 1);
 }
 
 void initRoom(rtype::ecs::Registry& reg)
